pyproject.toml
[project]
name = "raffle-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
license = "MIT"
authors = [
  { name = "Andoresu Peresu", email = "code@aperezc.xyz" }
]
maintainers = [
  { name = "Andoresu Peresu", email = "code@aperezc.xyz" }
]
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.10.6",
    "pyyaml>=6.0.2",
    "typer>=0.15.2",
]

[project.scripts]
raffle-cli = "raffle_cli:main_cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
